struct SensorId. Works for cartographer without grpc.
