+++ cartographer/mapping/collated_trajectory_builder.cc
@@ -30,14 +30,18 @@ constexpr double kSensorDataRatesLoggingPeriodSeconds = 15.;
 
 CollatedTrajectoryBuilder::CollatedTrajectoryBuilder(
     sensor::CollatorInterface* const sensor_collator, const int trajectory_id,
-    const std::unordered_set<SensorId>& expected_sensor_ids,
+    const std::set<SensorId>& expected_sensor_ids,
     std::unique_ptr<TrajectoryBuilderInterface> wrapped_trajectory_builder)
     : sensor_collator_(sensor_collator),
       trajectory_id_(trajectory_id),
       wrapped_trajectory_builder_(std::move(wrapped_trajectory_builder)),
       last_logging_time_(std::chrono::steady_clock::now()) {
+  std::unordered_set<std::string> expected_sensor_id_strings;
+  for (auto sensor_id : expected_sensor_ids) {
+    expected_sensor_id_strings.insert(sensor_id.id);
+  }
   sensor_collator_->AddTrajectory(
-      trajectory_id, expected_sensor_ids,
+      trajectory_id, expected_sensor_id_strings,
       [this](const std::string& sensor_id, std::unique_ptr<sensor::Data> data) {
         HandleCollatedSensorData(sensor_id, std::move(data));
       });
