+++ cartographer/mapping/map_builder.h
@@ -20,7 +20,7 @@
 #include "cartographer/mapping/map_builder_interface.h"
 
 #include <memory>
-#include <unordered_map>
+#include <set>
 
 #include "cartographer/common/thread_pool.h"
 #include "cartographer/mapping/pose_graph_interface.h"
@@ -46,7 +46,7 @@ class MapBuilder : public MapBuilderInterface {
   MapBuilder& operator=(const MapBuilder&) = delete;
 
   int AddTrajectoryBuilder(
-      const std::unordered_set<sensor::SensorId>& expected_sensor_ids,
+      const std::set<SensorId>& expected_sensor_ids,
       const proto::TrajectoryBuilderOptions& trajectory_options,
       LocalSlamResultCallback local_slam_result_callback) override;
