+++ cartographer/mapping/trajectory_builder_interface.h
@@ -19,6 +19,7 @@
 
 #include <functional>
 #include <memory>
+#include <set>
 #include <string>
 
 #include "cartographer/common/lua_parameter_dictionary.h"
@@ -61,6 +62,32 @@ class TrajectoryBuilderInterface {
                          transform::Rigid3d /* local pose estimate */,
                          sensor::RangeData /* in local frame */,
                          std::unique_ptr<const InsertionResult>)>;
+
+  struct SensorId {
+    enum class SensorType {
+      RANGE = 0,
+      IMU,
+      ODOMETRY,
+      FIXED_FRAME_POSE,
+      LANDMARK,
+      LOCAL_SLAM_RESULT
+    };
+
+    SensorType type;
+    std::string id;
+
+    bool operator==(const SensorId& other) const {
+      return std::forward_as_tuple(type, id) ==
+          std::forward_as_tuple(other.type, other.id);
+    }
+
+    bool operator!=(const SensorId& other) const { return !operator==(other); }
+
+    bool operator<(const SensorId& other) const {
+      return std::forward_as_tuple(type, id) <
+          std::forward_as_tuple(other.type, other.id);
+    }
+  };
 
   TrajectoryBuilderInterface() {}
   virtual ~TrajectoryBuilderInterface() {}
