--- cartographer/sensor/id.h
@@ -1,40 +1,0 @@
-/*
- * Copyright 2018 The Cartographer Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef CARTOGRAPHER_SENSOR_ID_H_
-#define CARTOGRAPHER_SENSOR_ID_H_
-
-namespace cartographer {
-namespace sensor {
-
-struct SensorId {
-  enum class SensorType {
-    RANGE = 0,
-    IMU,
-    ODOMETRY,
-    FIXED_FRAME_POSE,
-    LANDMARK,
-    LOCAL_SLAM_RESULT
-  };
-
-  SensorType type;
-  std::string id;
-};
-
-}  // namespace sensor
-}  // namespace cartographer
-
-#endif  // CARTOGRAPHER_SENSOR_ID_H_
